requirements.txt

python-telegram-bot==20.7
openai-whisper==20250625
psycopg2-binary==2.9.11
python-dotenv==1.2.1
